Hygiene: extract app initialization into separate loaders
This is mainly a repackaging of code into separate modules. In this
patch I've tried to preserve the functions as much as possible.

I try to keep the usual template config loading in commonConfig.ts.
We might delete a few items we don't need there but the main parts that
are specific to a particular service can go into config.ts.

Eventually I'd like to get to something close to what's described in
https://softwareontheroad.com/ideal-nodejs-project-structure/#loaders.

Change-Id: Icbbf4de7eb532a44fe1207f8d79ba30e13428ff3