Add self-updater script
Adds a script to check the git master for updates to the repo, and pull
them and restart the service if any are found.

This will be run on a cron job on the Cloud VPS testing instance,
as described in:
https://www.mediawiki.org/wiki/Wikimedia_Apps/Team/RESTBase_services_for_apps/Testing_instance_setup#Enable_automatic_updating.

This is based on a similar script existing in both the mobileapps and
wikifeeds projects at scripts/puller.sh. The name is updated here to
self_update.sh to be more descriptive.

Bug: T253647
Change-Id: I2ecd071b987fc0330eb69faa0c0d73b9ca733958