Add API rate limiting support
The push notifications request API will be exposed publicly in Cloud
VPS, both in the Beta Cluster and in any standalone instances we create.
The API should be rate limited to prevent abuse.

This patch adds support for rate limiting using the express-rate-limit
package. Rate limiting settings can be configured as usual in the app
config. A new config.beta.yaml configuration file is added, with strict
settings for Cloud VPS environments.

Change-Id: Ib85b4c213634996c7e67214463cc3543d88c8b21