Set GOOGLE_APPLICATION_CREDENTIALS in the production Docker image
The firebase-admin library expects the path to a credentials file to be
set in the GOOGLE_APPLICATION_CREDENTIALS environment variable. This
sets GOOGLE_APPLICATION_CREDENTIALS to
/etc/mediawiki-services-push-notifications/google-application-credentials.json.

Note: The /etc/mediawiki-services-push-notifications directory is
already copied over to the container when running the service in the
Beta Cluster and (presumably) production, as it's where the service
configuration file (config.yaml) lives.

This is needed to enable sending notification requests to FCM from
service running in the Beta Cluster.

Hygiene: Standardized some curly-bracket spacing.

Bug: T254164
Change-Id: I38262cf12760ef3b5352a5d27eed59ef1459ea6a