FCM: Provide for sending a message to multiple devices
Updates the v1/message/fcm endpoint contract to accept, rather than a
single deviceToken string, a deviceTokens key in which to specify an
array of device tokens to which to send the message. This allows clients
(i.e., MediaWiki) to more efficiently send notifications to users who
have registered tokens for multiple devices (e.g., a phone and a tablet).

Bug: T254359
Change-Id: I48498eac16ba460b0657d9799e01c7fe22631e73