Add request queueing and batch processing
Adds a queueing module, the purpose of which is to provide means by
which to increase the difficulty of inferring connections between
on-wiki actions and traffic on the public internet, in order to
mitigate user privacy risks associated with push notifications. It
does so by providing for batching requests up to a configurable max
count and/or length of time before submission. Further, the max
length of time for batching may be randomized within configurable
bounds.

Note that randomization will apply on a per-worker basis; in other
words, if queue flush timeouts are randomized, each worker node will
have a different random flush timeout.

Hygiene: Update some existing code to use async/await syntax. Mixing
async/await in queueing.ts and legacy promise syntax in other modules
caused difficult-to-reason-about failures when enqueueing or sending
APNS requests.

Bug: T255605
Change-Id: Ifa038b2d7197baa5f2cc86bb8adc585052aee375