apns/fcm: Implement proxy configuration
1) pass proxy object as configuration to apns and fcm libraries
2) move to upstream node-apn librarie and stop using HTTP/2 because of
the lack of proxy support

It seems that service-runner doesn't support proxy out of the box, this
patch introduces proxy support for apns and fcm libraries.

In order to test it, you need a proxy server. I used tinyproxy and you
can use the docker-compose environment I created for it in the following
link [1]

After that, you need to uncomment lines 47-49 in the config.dev.yaml
file. Test the service endpoints and look at the proxy log to see if
the requests are being processed by it. If the proxy configuration is
defined, and the proxy server is down, all requests made by fcm and apns
will fail.

[1] https://github.com/thesocialdev/simproxy

Bug: T256365
Bug: T256366
Change-Id: I888e5a21c8e854f5fe4a248dc1a8342af5f7dce0