Clean up invalid subscriptions
Upon one or more subscriptions being flagged as invalid by a provider
API response, makes a request to the MW API to delete them.

This requires some restructuring of the code in order to get access
to the application configuration when handling provider API responses.
Handlers for APNS, FCM, and MW API are restructured as Client objects
and attached to the application object on startup.

Note: We only expect to need to interact with one project
(meta.wikimedia.org), so that domain is used specifically in the MW
API request template.

This patch also fixes a bug in which APNS topics were not taken into
account when batching messages.

Bug: T260247
Change-Id: Ic50da23372b04e024662b8b5152967d8e4b0935f