FCM: Pass http.Agent to the Credential factory on initializing app
Outgoing OAuth token fetch requests are currently failing in
production because they are not honoring the proxy configuration.
This is because the Credential does not automatically use the same
http.Agent as the one passed in the AppOptions (if any).

This updates the call to admin.initializeApp() to also pass the
configured ProxyAgent to the Credential factory.

Bug: T263695
Change-Id: Idd51fe2c62ad69e65b6666a5beb39e2985357743