Throw on MW API error response
Currently the mwApiPost method (inherited as mwApiGet from service-
template-node) incorrectly handles MW API errors. It checks for
response codes < 200 or > 399, but this is incorrect because MW API
error responses have status 200 and are only detectable as errors
from the response body.

This updates the API error handling to throw when the API response
conforms to the standard error structure:

 {
   "error": {
     "code": "errorcode",
     "info": "Error message text"
   }
 }

Change-Id: I6aa32cb2ad7198583ef5ea205030b5e01c059a01